change: 숨겨진 공간 렌더링 개선

+++ src/options/game_option.rs
@@ -3,8 +3,8 @@ use crate::game::bag::BagType;
 pub struct GameOption {
     pub board_width: u32,
     pub board_height: u32,
-    pub column_count: u8,
-    pub row_count: u8,
+    pub column_count: u32,
+    pub row_count: u32,
     pub bag_mode: BagType,
 }
